requirements.txt
more_itertools
numpy
tensorboardX
torch
tqdm
pytest
pyyaml
wandb
git+https://github.com/Hiroshiba/pytorch-trainer@1.3.1
git+https://github.com/Hiroshiba/acoustic_feature_extractor
